# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "E-commerce REST API with products, categories, orders, users, and inventory management",
        "title": "Bey API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@bey.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/payments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new payment transaction through Wompi gateway for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create a new payment",
                "parameters": [
                    {
                        "description": "Payment creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_payments.CreatePaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Payment created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_payments.PaymentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid payment data",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - payment processing failed",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/links": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new payment link through Wompi gateway that can be shared with customers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Create a payment link",
                "parameters": [
                    {
                        "description": "Payment link creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_payments.CreatePaymentLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Payment link created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_payments.PaymentLinkResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid payment link data",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - payment link creation failed",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/links/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves details of a specific payment link by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment link details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment link ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment link retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_payments.PaymentLinkResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid payment link ID",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Payment link not found",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/links/{id}/activate": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Activates a previously deactivated payment link",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Activate a payment link",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment link ID to activate",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment link activated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_payments.PaymentLinkResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - payment link cannot be activated",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Payment link not found",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/links/{id}/deactivate": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deactivates an active payment link to prevent further payments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Deactivate a payment link",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Payment link ID to deactivate",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment link deactivated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_payments.PaymentLinkResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - payment link cannot be deactivated",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Payment link not found",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/webhook": {
            "post": {
                "description": "Receives and processes webhook events from Wompi payment gateway",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Handle Wompi webhook",
                "parameters": [
                    {
                        "description": "Webhook event payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_payments.WebhookEvent"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook processed successfully",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid payload",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid webhook signature",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error - webhook processing failed",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/payments/wompi/{wompi_id}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves payment status using Wompi transaction ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment status by Wompi ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wompi transaction ID",
                        "name": "wompi_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment status retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_modules_payments.PaymentResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid Wompi ID or payment error",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/bey_internal_shared_response.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bey_internal_shared_response.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "internal_modules_payments.CreatePaymentRequest": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "payment_token",
                "reference"
            ],
            "properties": {
                "amount": {
                    "description": "Amount in cents",
                    "type": "integer"
                },
                "currency": {
                    "description": "e.g., \"COP\"",
                    "type": "string"
                },
                "payment_token": {
                    "description": "Wompi payment token",
                    "type": "string"
                },
                "redirect_url": {
                    "description": "Optional redirect after payment",
                    "type": "string"
                },
                "reference": {
                    "description": "Order reference",
                    "type": "string"
                }
            }
        },
...
        "internal_modules_payments.PaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "order_id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string"
                },
                "redirect_url": {
                    "type": "string"
                },
                "reference": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "wompi_transaction_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your Bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}